create_and_save_monomer_objects made readable. Some functions moved to utils
